+++ src/components/room/RoomFooter.tsx
@@ -1,5 +1,5 @@
 import dayjs from 'dayjs';
-import { FormEvent, KeyboardEvent, useEffect, useRef } from 'react';
+import { KeyboardEvent, useEffect, useRef } from 'react';
 import styled from 'styled-components';
 import useInput from '../../hooks/useInput';
 import { IChat } from '../../states/interface';
